Width/height sanity check

+++ coders/miff.c
@@ -233,7 +233,8 @@ static void PushRunlengthPacket(Image *image,const unsigned char *pixels,
         }
         case 16:
         {
-          *index=ConstrainColormapIndex(image,(ssize_t) ((*p << 8) | *(p+1)));
+          *index=ConstrainColormapIndex(image,(ssize_t) ((size_t) (*p << 8) |
+            (size_t) *(p+1)));
           p+=2;
           break;
         }
@@ -292,7 +293,7 @@ static void PushRunlengthPacket(Image *image,const unsigned char *pixels,
           (void) ThrowMagickException(&image->exception,GetMagickModule(),
             CorruptImageError,"ImageDepthNotSupported","`%s'",image->filename);
       }
-      *length=(size_t) (*p++)+1;
+      *length=((size_t) *p++)+1;
       return;
     }
   switch (image->depth)
